move OS definitions to config
