+++ common/v2/config/index.ts
@@ -28,3 +28,4 @@ export { accounts as testAccounts } from './accounts';
 export { assets as testAssets } from './assets';
 export { addressBook as testAddressBook } from './addressBook';
 export { settings as testSettings } from './settings';
+export { OS } from './operatingSystems';
