+++ common/v2/config/operatingSystems.ts
@@ -1,0 +1,7 @@
+export enum OS {
+  WINDOWS = 'windows',
+  MAC = 'mac',
+  LINUX64 = 'linux64',
+  LINUX32 = 'linux32',
+  STANDALONE = 'standalone'
+}
