+++ common/v2/features/DownloadApp/DownloadApp.tsx
@@ -5,8 +5,8 @@ import cloneDeep from 'lodash/cloneDeep';
 import styled from 'styled-components';
 
 import { ExtendedContentPanel } from 'v2/components';
-import { AnalyticsService, ANALYTICS_CATEGORIES, GithubService, OS } from 'v2/services/ApiService';
-import { GITHUB_RELEASE_NOTES_URL, DOWNLOAD_MYCRYPTO_LINK } from 'v2/config';
+import { AnalyticsService, ANALYTICS_CATEGORIES, GithubService } from 'v2/services/ApiService';
+import { GITHUB_RELEASE_NOTES_URL, DOWNLOAD_MYCRYPTO_LINK, OS } from 'v2/config';
 import { getFeaturedOS } from 'v2/utils';
 import { AppDownloadItem } from './types';
 import translate from 'translations';
