+++ common/v2/services/ApiService/Github/constants.ts
@@ -1,9 +1,1 @@
 export const GITHUB_RELEASES_URL = 'https://github.proxy.mycryptoapi.com/';
-
-export enum OS {
-  WINDOWS = 'windows',
-  MAC = 'mac',
-  LINUX64 = 'linux64',
-  LINUX32 = 'linux32',
-  STANDALONE = 'standalone'
-}
