+++ common/v2/services/ApiService/Github/index.ts
@@ -1,3 +1,2 @@
 export { default as GithubService } from './Github';
 export * from './types';
-export { OS } from './constants';
