+++ common/v2/services/ApiService/index.ts
@@ -2,5 +2,5 @@ export { default as ApiService } from './ApiService';
 export { AnalyticsService, ANALYTICS_CATEGORIES, Params, CvarEntry } from './Analytics';
 export { subscribeToMailingList } from './emails';
 export { getDefaultEstimates, fetchGasPriceEstimates, getGasEstimate } from './Gas';
-export { GithubService, OS } from './Github';
+export { GithubService } from './Github';
 export * from './constants';
